requirements.txt
matplotlib==3.5.1
numpy==1.21.6
opencv-python==4.8.1.78
pandas==1.3.5
pytube==15.0.0
PyYAML==5.4.1
scikit-learn==1.0.1
tqdm==4.62.3